+++ src/parser/DataSource.java
@@ -6,8 +6,8 @@ public interface DataSource
 {	
 	public void read(String directory, ParserFactory parser);
 	
-	public Map<Date, Integer> getAbsoluteOccupancy(String labName);	
-	public Map<Date, Float> getRelativeOccupancy(String labName);
+	public Map<Date, Double> getAbsoluteOccupancy(String labName);	
+	public Map<Date, Double> getRelativeOccupancy(String labName);
 	
 	public List<String> getLabList();	
 }
